TELE/requirements.txt
pandas>=1.5.3
numpy>=1.24.0
sentence-transformers>=2.2.2
torch>=2.1.0
python-telegram-bot>=20.3
python-dotenv>=1.0.0
